Combine the two tempest thrift services (which had been separate for historical reasons) and re-organize methods.
